requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-allauth==0.51.0
django-celery-beat==2.3.0
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
postgres==4.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
